Initial commit for the new updates to proteus.
The following are the major changes here:
  - native code implementation
  - unit test infrastructure started
  - unity build systems established

The initial goal for proteus is to be a cleaner c-style language, allowing certain constructs
that are challenging to implement in C without being overly verbose.

There are **no** fundamental changes to the C-language being done, such as the memory model as
is done in Rust. The first stage of this is simply a proteus->c transpiler.